+++ beetsplug/bpd/__init__.py
@@ -27,7 +27,10 @@ from beets.util import as_string
 from beetsplug._utils import vfs
 
 if TYPE_CHECKING:
+    import optparse
+
     from beets.dbcore.query import Query
+    from beets.library import Library
 
 
 try:
@@ -1621,21 +1624,19 @@ class BPDPlugin(BeetsPlugin):
             "bpd", help="run an MPD-compatible music player server"
         )
 
-        def func(lib, opts, args):
+        def func(lib: Library, opts: optparse.Values, args: list[str]) -> None:
             host = self.config["host"].as_str()
             host = args.pop(0) if args else host
             port = args.pop(0) if args else self.config["port"].get(int)
             if args:
-                ctrl_port = args.pop(0)
+                ctrl_port = int(args.pop(0))
             else:
                 ctrl_port = self.config["control_port"].get(int)
             if args:
                 raise UserError("too many arguments")
             password = self.config["password"].as_str()
             volume = self.config["volume"].get(int)
-            self.start_bpd(
-                lib, host, int(port), password, volume, int(ctrl_port)
-            )
+            self.start_bpd(lib, host, int(port), password, volume, ctrl_port)
 
         cmd.func = func
         return [cmd]