+++ beetsplug/play.py
@@ -6,7 +6,7 @@ import random
 import shlex
 import subprocess
 from os.path import relpath
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Protocol
 
 from beets import config, ui, util
 from beets.exceptions import UserError
@@ -16,12 +16,23 @@ from beets.util import PromptChoice, get_temp_filename
 from beets.util.color import colorize
 
 if TYPE_CHECKING:
+    from collections.abc import Sequence
+
     from beets.importer import ImportSession, ImportTask
+    from beets.library import LibModel, Library
 
 
 # Indicate where arguments should be inserted into the command string.
 # If this is missing, they're placed at the end.
 ARGS_MARKER = "$args"
+
+
+class PlayCLIOpts(Protocol):
+    album: bool
+    args: str | None
+    randomize: bool | None
+    yes: bool | None
+
 
 # Indicate where the playlist file (with absolute path) should be inserted into
 # the command string. If this is missing, its placed at the end, but before
@@ -102,7 +113,9 @@ class PlayPlugin(BeetsPlugin):
         play_command.func = self._play_command
         return [play_command]
 
-    def _play_command(self, lib, opts, args):
+    def _play_command(
+        self, lib: Library, opts: PlayCLIOpts, args: list[str]
+    ) -> None:
         """The CLI command function for `beet play`. Create a list of paths
         from query, determine if tracks or albums are to be played.
         """
@@ -112,6 +125,7 @@ class PlayPlugin(BeetsPlugin):
             relative_to = util.normpath(relative_to)
         # Perform search by album and add folders rather than tracks to
         # playlist.
+        selection: Sequence[LibModel]
         if opts.album:
             selection = lib.albums(args)
             paths = []