+++ beetsplug/web/__init__.py
@@ -1,9 +1,12 @@
 """A Web interface to beets."""
 
+from __future__ import annotations
+
 import base64
 import json
 import os
 import typing as t
+from typing import TYPE_CHECKING, Protocol
 
 import flask
 from flask import jsonify
@@ -14,6 +17,14 @@ import beets.library
 from beets import ui, util
 from beets.dbcore.query import PathQuery
 from beets.plugins import BeetsPlugin
+
+if TYPE_CHECKING:
+    from beets.library import Library
+
+
+class WebCLIOpts(Protocol):
+    debug: bool
+
 
 # Type checking hacks
 
@@ -445,7 +456,7 @@ class WebPlugin(BeetsPlugin):
             help="debug mode",
         )
 
-        def func(lib, opts, args):
+        def func(lib: Library, opts: WebCLIOpts, args: list[str]) -> None:
             args = args
             if args:
                 self.config["host"] = args.pop(0)
@@ -479,7 +490,7 @@ class WebPlugin(BeetsPlugin):
 
             # Allow serving behind a reverse proxy
             if self.config["reverse_proxy"]:
-                app.wsgi_app = ReverseProxied(app.wsgi_app)
+                app.wsgi_app = ReverseProxied(app.wsgi_app)  # type: ignore[method-assign]
 
             # Start the web application.
             app.run(